@@ -10,9 +10,9 @@
         }
     ],
     "require": {
-        "php": "^7.3 || ^8.0",
+        "php": "^7.4 || ^8.0",
         "ext-tokenizer": "*",
-        "friendsofphp/php-cs-fixer": "^3.1"
+        "friendsofphp/php-cs-fixer": "3.4"
     },
     "conflict": {
         "liaison/cs-config": "*"
